added filtering data for categories and added binding todos/categories to user

+++ ToDoList/Controllers/CategoriesController.cs
@@ -1,5 +1,7 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using ToDoList.Enums;
+using ToDoList.Services;
 
 namespace ToDoList.Controllers
 {
@@ -14,9 +16,10 @@ namespace ToDoList.Controllers
         }
 
         // GET: CategoriesController
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(string? like, CategoriesSortOrder sortOrder)
         {
-            return View(await _categoryRepository.GetAsync());
+            int currentUserId = int.Parse(HttpContext.User.Claims.First(c => c.Type == AccountService.DefaultIdClaimType).Value);
+            return View(await _categoryRepository.GetMyAsync(currentUserId, like, sortOrder));
         }
 
         // GET: CategoriesController/Details/5
@@ -25,6 +28,14 @@ namespace ToDoList.Controllers
             CategoryModel category = await _categoryRepository.GetByIdWithTodosAsync(id);
             if(category == null)
                 return NotFound();
+
+            int currentUserId = int.Parse(HttpContext.User.Claims.First(c => c.Type == AccountService.DefaultIdClaimType).Value);
+            if (category.UserId != currentUserId)
+            {
+                TempData["Error"] = "You dont have access to details it.";
+                return RedirectToAction(nameof(Index));
+            }
+
             return View(category);
         }
 
@@ -41,6 +52,7 @@ namespace ToDoList.Controllers
         {
             if (ModelState.IsValid)
             {
+                category.UserId = int.Parse(HttpContext.User.Claims.First(c => c.Type == AccountService.DefaultIdClaimType).Value);
                 await _categoryRepository.CreateAsync(category);
                 return RedirectToAction(nameof(Index));
             }
@@ -53,6 +65,14 @@ namespace ToDoList.Controllers
             CategoryModel category = await _categoryRepository.GetByIdAsync(id);
             if (category == null)
                 return NotFound();
+
+            int currentUserId = int.Parse(HttpContext.User.Claims.First(c => c.Type == AccountService.DefaultIdClaimType).Value);
+            if (category.UserId != currentUserId)
+            {
+                TempData["Error"] = "You dont have access to edit it.";
+                return RedirectToAction(nameof(Index));
+            }
+
             return View(category);
         }
 
@@ -63,16 +83,24 @@ namespace ToDoList.Controllers
         {
             if (ModelState.IsValid)
             {
-                try
+                int currentUserId = int.Parse(HttpContext.User.Claims.First(c => c.Type == AccountService.DefaultIdClaimType).Value);
+                var checkAccessCategory = await _categoryRepository.GetByIdAsync(category.Id);
+                if (checkAccessCategory.UserId != currentUserId)
                 {
-                    await _categoryRepository.UpdateAsync(category);
+                    ModelState.AddModelError("", "You dont have access to edit it");
                 }
-                catch(Exception ex)
+                else
                 {
-                    ModelState.AddModelError("", ex.Message);
-                    return View(category);
+                    try
+                    {
+                        await _categoryRepository.UpdateAsync(category);
+                        return RedirectToAction(nameof(Index));
+                    }
+                    catch (Exception ex)
+                    {
+                        ModelState.AddModelError("", ex.Message);
+                    }
                 }
-                return RedirectToAction(nameof(Index));
             }
             return View(category);
         }
@@ -83,6 +111,14 @@ namespace ToDoList.Controllers
             CategoryModel category = await _categoryRepository.GetByIdAsync(id);
             if (category == null)
                 return NotFound();
+
+            int currentUserId = int.Parse(HttpContext.User.Claims.First(c => c.Type == AccountService.DefaultIdClaimType).Value);
+            if (category.UserId != currentUserId)
+            {
+                TempData["Error"] = "You dont have access to delete it.";
+                return RedirectToAction(nameof(Index));
+            }
+
             return View(category);
         }
 
@@ -94,6 +130,14 @@ namespace ToDoList.Controllers
             CategoryModel category = await _categoryRepository.GetByIdAsync(id);
             if (category == null)
                 return NotFound();
+
+            int currentUserId = int.Parse(HttpContext.User.Claims.First(c => c.Type == AccountService.DefaultIdClaimType).Value);
+            if (category.UserId != currentUserId)
+            {
+                TempData["Error"] = "You dont have access to delete it.";
+                return View(category);
+            }
+
             await _categoryRepository.RemoveAsync(category.Id);
             return RedirectToAction(nameof(Index));
         }
